finished refactoring
