+++ CMakeLists.txt
@@ -12,11 +12,11 @@ set(CMAKE_EXPORT_COMPILE_COMMANDS ON)
 
 # Project/Library Names
 set(CMAKE_MODULE_PATH ${PROJECT_SOURCE_DIR}/cmake/)
-set(EXECUTABLE_NAME Executable)
-set(LIBRARY_MODEL_NAME model)
-set(LIBRARY_PRESENTER_NAME Presenter)
-set(LIBRARY_VIEW_NAME view)
+set(EXECUTABLE_NAME serialLogger)
+set(LIBRARY_GUI_BACKEND_NAME guiBackend)
+set(LIBRARY_IGUI_BACKEND_NAME I${LIBRARY_GUI_BACKEND_NAME})
 set(LIBRARY_NET_NAME Netlib)
+set(LIBRARY_INET_NAME I${LIBRARY_NET_NAME})
 set(LIBRARY_NETSERVER_NAME Server)
 
 # Options
