+++ app/CMakeLists.txt
@@ -1,27 +1,23 @@
 set(EXE_SOURCES
-    "serial-logger.cpp")
+    "serialLogger.cpp")
 
 add_executable(${EXECUTABLE_NAME} ${EXE_SOURCES})
 
 target_link_libraries(${EXECUTABLE_NAME} PRIVATE 
-${LIBRARY_MODEL_NAME}
-${LIBRARY_VIEW_NAME}
-${LIBRARY_PRESENTER_NAME}
-${LIBRARY_NET_NAME}
-${LIBRARY_NETSERVER_NAME}
-glfw
-OpenGL::GL
-setupapi
-ws2_32
-imgui::imgui
-implot::implot
+    ${LIBRARY_NETSERVER_NAME}
+    ${LIBRARY_GUI_BACKEND_NAME}
+    ${LIBRARY_NETSERVER_NAME}
+    glfw
+    OpenGL::GL
+    setupapi
+    ws2_32
+    imgui::imgui
+    implot::implot
 )
 
 if(ENABLE_CLANG_FORMAT)
     add_dependencies(${EXECUTABLE_NAME} run_clang_format)
 endif()
-
-# add_dependencies(${EXECUTABLE_NAME} ${LIBRARY_VIEW_NAME})
 
 if (ENABLE_WARNINGS)
     target_set_warnings(TARGET ${EXECUTABLE_NAME}
