+++ modules/CMakeLists.txt
@@ -1,4 +1,2 @@
-add_subdirectory(model ./Model)
-add_subdirectory(Presenter ./Presenter)
-add_subdirectory(view ./View)
-add_subdirectory(SerialPort ./SerialPort)
+add_subdirectory(guiBackend ./guiBackend)
+add_subdirectory(serialPort ./serialPort)
