+++ modules/Serialport/Netlib/interfaces/connection/Iconnection.hpp
