+++ modules/Serialport/Netlib/interfaces/serverBase/IserverBase.hpp
@@ -1,7 +1,7 @@
 #pragma once
 #include <memory>
 #include <string>
-#include <TSQueue/Message.hpp>
+#include <tsQueue/types/message.hpp>
 
 namespace netlib
 {
