+++ modules/Serialport/Server/CMakeLists.txt
@@ -1,5 +1,5 @@
 set(SOURCES 
-    "Server.cpp")
+    "./src/server.cpp")
 
 add_library(${LIBRARY_NETSERVER_NAME} STATIC 
     ${SOURCES})
@@ -8,14 +8,14 @@ target_include_directories(${LIBRARY_NETSERVER_NAME} PUBLIC
     ${CMAKE_CURRENT_SOURCE_DIR}/include)
 
 target_link_libraries(${LIBRARY_NETSERVER_NAME} PUBLIC
+    ${LIBRARY_IGUI_BACKEND_NAME}
+    ${LIBRARY_INET_NAME}
+    )
+
+target_link_libraries(${LIBRARY_NETSERVER_NAME} PRIVATE
     ${LIBRARY_NET_NAME}
-    ${LIBRARY_MODEL_NAME}
-    Boost::system
-    Boost::asio)
-
-#Use the following when switching to shared library
-#target_compile_definitions(${LIBRARY_VIEW_NAME} PRIVATE BUILDING_VIEW_DLL)
-
+    )
+    
 
 if (ENABLE_WARNINGS)
     target_set_warnings(TARGET ${LIBRARY_NETSERVER_NAME}
