+++ modules/guiBackend/interfaces/CMakeLists.txt
@@ -1,0 +1,7 @@
+add_library(${LIBRARY_IGUI_BACKEND_NAME} INTERFACE)
+
+target_include_directories(${LIBRARY_IGUI_BACKEND_NAME} INTERFACE ${CMAKE_CURRENT_SOURCE_DIR})
+
+target_link_libraries(${LIBRARY_IGUI_BACKEND_NAME} INTERFACE
+    ${LIBRARY_NET_NAME}
+)
