+++ modules/guiBackend/src/presenter.cpp
@@ -1,4 +1,4 @@
-#include <Presenter/Presenter.hpp>
+#include <presenter/presenter.hpp>
 
 void Presenter::start()
 {
